chore: move folder to plugins folder
