+++ src/plugins/university/deps.ts
