+++ src/plugins/university/src/Bitfields/Bitfield.ts
