+++ src/plugins/university/src/Bitfields/User.ts
