+++ src/plugins/university/src/Client.ts
@@ -1,14 +1,14 @@
-import { BotConfig } from "../../bot.ts";
-import { PresenceUpdate } from "../../types/activity/presence_update.ts";
-import { DiscordGatewayOpcodes } from "../../types/codes/gateway_opcodes.ts";
-import { Errors } from "../../types/discordeno/errors.ts";
-import { Intents } from "../../types/gateway/gateway_intents.ts";
-import { StatusUpdate } from "../../types/gateway/status_update.ts";
-import { User } from "../../types/users/user.ts";
-import { snowflakeToBigint } from "../../util/bigint.ts";
-import { Collection } from "../../util/collection.ts";
-import { endpoints } from "../../util/constants.ts";
-import { snakelize, urlToBase64 } from "../../util/utils.ts";
+import { BotConfig } from "../../../bot.ts";
+import { PresenceUpdate } from "../../../types/activity/presence_update.ts";
+import { DiscordGatewayOpcodes } from "../../../types/codes/gateway_opcodes.ts";
+import { Errors } from "../../../types/discordeno/errors.ts";
+import { Intents } from "../../../types/gateway/gateway_intents.ts";
+import { StatusUpdate } from "../../../types/gateway/status_update.ts";
+import { User } from "../../../types/users/user.ts";
+import { snowflakeToBigint } from "../../../util/bigint.ts";
+import { Collection } from "../../../util/collection.ts";
+import { endpoints } from "../../../util/constants.ts";
+import { snakelize, urlToBase64 } from "../../../util/utils.ts";
 import { EventEmitter, decode } from "../deps.ts";
 import Channel from "./Channel.ts";
 import { GatewayManager } from "./Gateway/GatewayManager.ts";
