+++ src/plugins/university/src/Gateway/Events.ts
@@ -1,50 +1,50 @@
-import { PresenceUpdate } from "../../../types/activity/presence_update.ts";
-import { Channel as ChannelPayload } from "../../../types/channels/channel.ts";
-import { ChannelPinsUpdate } from "../../../types/channels/channel_pins_update.ts";
-import { DiscordChannelTypes } from "../../../types/channels/channel_types.ts";
-import { StageInstance } from "../../../types/channels/stage_instance.ts";
-import { ThreadListSync } from "../../../types/channels/threads/thread_list_sync.ts";
-import { ThreadMember } from "../../../types/channels/threads/thread_member.ts";
-import { ThreadMembersUpdate } from "../../../types/channels/threads/thread_members_update.ts";
-import { GuildUpdateChange } from "../../../types/discordeno/guild_update_change.ts";
-import { GuildEmojisUpdate } from "../../../types/emojis/guild_emojis_update.ts";
-import { DiscordGatewayPayload } from "../../../types/gateway/gateway_payload.ts";
-import { Ready } from "../../../types/gateway/ready.ts";
-import { Guild as GuildPayload } from "../../../types/guilds/guild.ts";
-import { GuildBanAddRemove } from "../../../types/guilds/guild_ban_add_remove.ts";
-import { GuildRoleCreate } from "../../../types/guilds/guild_role_create.ts";
-import { GuildRoleDelete } from "../../../types/guilds/guild_role_delete.ts";
-import { GuildRoleUpdate } from "../../../types/guilds/guild_role_update.ts";
-import { UnavailableGuild } from "../../../types/guilds/unavailable_guild.ts";
-import { GuildIntegrationsUpdate } from "../../../types/integrations/guild_integrations_update.ts";
-import { IntegrationCreateUpdate } from "../../../types/integrations/integration_create_update.ts";
-import { IntegrationDelete } from "../../../types/integrations/integration_delete.ts";
-import { ApplicationCommandCreateUpdateDelete } from "../../../types/interactions/commands/application_command_create_update_delete.ts";
-import { Interaction } from "../../../types/interactions/interaction.ts";
-import { DiscordInteractionTypes } from "../../../types/interactions/interaction_types.ts";
-import { InviteCreate } from "../../../types/invites/invite_create.ts";
-import { InviteDelete } from "../../../types/invites/invite_delete.ts";
-import { GuildMemberWithUser } from "../../../types/members/guild_member.ts";
-import { GuildMembersChunk } from "../../../types/members/guild_members_chunk.ts";
-import { GuildMemberAdd } from "../../../types/members/guild_member_add.ts";
-import { GuildMemberRemove } from "../../../types/members/guild_member_remove.ts";
-import { GuildMemberUpdate } from "../../../types/members/guild_member_update.ts";
-import { MessageComponentTypes } from "../../../types/messages/components/message_component_types.ts";
-import { Message as MessagePayload } from "../../../types/messages/message.ts";
-import { MessageDelete } from "../../../types/messages/message_delete.ts";
-import { MessageDeleteBulk } from "../../../types/messages/message_delete_bulk.ts";
-import { MessageReactionAdd } from "../../../types/messages/message_reaction_add.ts";
-import { MessageReactionRemove } from "../../../types/messages/message_reaction_remove.ts";
-import { MessageReactionRemoveAll } from "../../../types/messages/message_reaction_remove_all.ts";
-import { MessageReactionRemoveEmoji } from "../../../types/messages/message_reaction_remove_emoji.ts";
-import { TypingStart } from "../../../types/misc/typing_start.ts";
-import { User } from "../../../types/users/user.ts";
-import { VoiceServerUpdate } from "../../../types/voice/voice_server_update.ts";
-import { VoiceState as VoiceStatePayload } from "../../../types/voice/voice_state.ts";
-import { WebhookUpdate } from "../../../types/webhooks/webhooks_update.ts";
-import { bigintToSnowflake, snowflakeToBigint } from "../../../util/bigint.ts";
-import { Collection } from "../../../util/collection.ts";
-import { iconHashToBigInt } from "../../../util/hash.ts";
+import { PresenceUpdate } from "../../../../types/activity/presence_update.ts";
+import { Channel as ChannelPayload } from "../../../../types/channels/channel.ts";
+import { ChannelPinsUpdate } from "../../../../types/channels/channel_pins_update.ts";
+import { DiscordChannelTypes } from "../../../../types/channels/channel_types.ts";
+import { StageInstance } from "../../../../types/channels/stage_instance.ts";
+import { ThreadListSync } from "../../../../types/channels/threads/thread_list_sync.ts";
+import { ThreadMember } from "../../../../types/channels/threads/thread_member.ts";
+import { ThreadMembersUpdate } from "../../../../types/channels/threads/thread_members_update.ts";
+import { GuildUpdateChange } from "../../../../types/discordeno/guild_update_change.ts";
+import { GuildEmojisUpdate } from "../../../../types/emojis/guild_emojis_update.ts";
+import { DiscordGatewayPayload } from "../../../../types/gateway/gateway_payload.ts";
+import { Ready } from "../../../../types/gateway/ready.ts";
+import { Guild as GuildPayload } from "../../../../types/guilds/guild.ts";
+import { GuildBanAddRemove } from "../../../../types/guilds/guild_ban_add_remove.ts";
+import { GuildRoleCreate } from "../../../../types/guilds/guild_role_create.ts";
+import { GuildRoleDelete } from "../../../../types/guilds/guild_role_delete.ts";
+import { GuildRoleUpdate } from "../../../../types/guilds/guild_role_update.ts";
+import { UnavailableGuild } from "../../../../types/guilds/unavailable_guild.ts";
+import { GuildIntegrationsUpdate } from "../../../../types/integrations/guild_integrations_update.ts";
+import { IntegrationCreateUpdate } from "../../../../types/integrations/integration_create_update.ts";
+import { IntegrationDelete } from "../../../../types/integrations/integration_delete.ts";
+import { ApplicationCommandCreateUpdateDelete } from "../../../../types/interactions/commands/application_command_create_update_delete.ts";
+import { Interaction } from "../../../../types/interactions/interaction.ts";
+import { DiscordInteractionTypes } from "../../../../types/interactions/interaction_types.ts";
+import { InviteCreate } from "../../../../types/invites/invite_create.ts";
+import { InviteDelete } from "../../../../types/invites/invite_delete.ts";
+import { GuildMemberWithUser } from "../../../../types/members/guild_member.ts";
+import { GuildMembersChunk } from "../../../../types/members/guild_members_chunk.ts";
+import { GuildMemberAdd } from "../../../../types/members/guild_member_add.ts";
+import { GuildMemberRemove } from "../../../../types/members/guild_member_remove.ts";
+import { GuildMemberUpdate } from "../../../../types/members/guild_member_update.ts";
+import { MessageComponentTypes } from "../../../../types/messages/components/message_component_types.ts";
+import { Message as MessagePayload } from "../../../../types/messages/message.ts";
+import { MessageDelete } from "../../../../types/messages/message_delete.ts";
+import { MessageDeleteBulk } from "../../../../types/messages/message_delete_bulk.ts";
+import { MessageReactionAdd } from "../../../../types/messages/message_reaction_add.ts";
+import { MessageReactionRemove } from "../../../../types/messages/message_reaction_remove.ts";
+import { MessageReactionRemoveAll } from "../../../../types/messages/message_reaction_remove_all.ts";
+import { MessageReactionRemoveEmoji } from "../../../../types/messages/message_reaction_remove_emoji.ts";
+import { TypingStart } from "../../../../types/misc/typing_start.ts";
+import { User } from "../../../../types/users/user.ts";
+import { VoiceServerUpdate } from "../../../../types/voice/voice_server_update.ts";
+import { VoiceState as VoiceStatePayload } from "../../../../types/voice/voice_state.ts";
+import { WebhookUpdate } from "../../../../types/webhooks/webhooks_update.ts";
+import { bigintToSnowflake, snowflakeToBigint } from "../../../../util/bigint.ts";
+import { Collection } from "../../../../util/collection.ts";
+import { iconHashToBigInt } from "../../../../util/hash.ts";
 import Channel from "../Channel.ts";
 import Client from "../Client.ts";
 import { Guild } from "../Guild.ts";
