+++ src/plugins/university/src/Guild.ts
@@ -1,23 +1,23 @@
 import Base from "./Base.ts";
 import Client from "./Client.ts";
 
-import { snowflakeToBigint } from "../../util/bigint.ts";
-import { Guild as GuildPayload } from "../../types/guilds/guild.ts";
-import { Collection } from "../../util/collection.ts";
+import { snowflakeToBigint } from "../../../util/bigint.ts";
+import { Guild as GuildPayload } from "../../../types/guilds/guild.ts";
+import { Collection } from "../../../util/collection.ts";
 import Channel from "./Channel.ts";
 import Member from "./Member.ts";
 import VoiceState from "./VoiceState.ts";
-import { Emoji } from "../../types/emojis/emoji.ts";
+import { Emoji } from "../../../types/emojis/emoji.ts";
 import Role from "./Role.ts";
-import { CreateGuildChannel, DiscordCreateGuildChannel } from "../../types/guilds/create_guild_channel.ts";
-import { endpoints } from "../../util/constants.ts";
-import { DiscordChannelTypes } from "../../types/channels/channel_types.ts";
-import { calculateBits } from "../../util/permissions.ts";
-import { hasOwnProperty, snakelize } from "../../util/utils.ts";
-import { UpdateSelfVoiceState } from "../../types/guilds/update_self_voice_state.ts";
-import { UpdateOthersVoiceState } from "../../types/guilds/update_others_voice_state.ts";
-import { ModifyGuildChannelPositions } from "../../types/guilds/modify_guild_channel_position.ts";
-import { Channel as ChannelPayload } from "../../types/channels/channel.ts";
+import { CreateGuildChannel, DiscordCreateGuildChannel } from "../../../types/guilds/create_guild_channel.ts";
+import { endpoints } from "../../../util/constants.ts";
+import { DiscordChannelTypes } from "../../../types/channels/channel_types.ts";
+import { calculateBits } from "../../../util/permissions.ts";
+import { hasOwnProperty, snakelize } from "../../../util/utils.ts";
+import { UpdateSelfVoiceState } from "../../../types/guilds/update_self_voice_state.ts";
+import { UpdateOthersVoiceState } from "../../../types/guilds/update_others_voice_state.ts";
+import { ModifyGuildChannelPositions } from "../../../types/guilds/modify_guild_channel_position.ts";
+import { Channel as ChannelPayload } from "../../../types/channels/channel.ts";
 
 export class Guild extends Base {
   /** The channels available in this guild. */
